requirements.txt
typer>=0.9.0
rich>=13.7.0
aider-chat[all]>=0.71.1
